Break out events filtering into new hook useEventsFilter
